+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes
 added;
 - `trmobjfunc` supports planar angles as variables for optimization;
 - Function `trmplotminatomdist` to plot minimal interatomic distances; 
+- Function `trmchangeangles` for convenient angle change in a model;
 - Option to leave only alpha carbon atoms added to `pdbbackbone`;
 - Option to consider planar angles added to `trmdistantangleindices`;
 - `trmcreate` creates a alpha carbon-based transformation model if the